+++ tools/vgsasm/tests/common.hpp
@@ -9,3 +9,5 @@
 #include "../src/decimal.hpp"
 #include "../src/string.hpp"
 #include "../src/tables.hpp"
+#include "../src/nametable.hpp"
+#include "../src/operand.hpp"